PR review comments: converter logic handler changes, removed unnecessary test code
